requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-pptx==1.0.2
openpyxl==3.1.5
google-generativeai==0.8.6
lxml==5.3.1
python-multipart==0.0.20
